Module Ocb_stubblr.Pkg_config

Query pkg-config.

pkg-config is invoked with the environment extended with the equivalent of

PKG_CONFIG_PATH=$(opam config var prefix)/{lib,share}/pkgconfig

. This means that any .pc files in the current Opam switch take precedence over the system-wide ones.

val run : flags:string list -> string -> [ `Nonexistent | `Res of string ]

run ~flags package queries pkg-config about package, using flags.

`Nonexistent means that the package was not found. Otherwise, `Res output is whatever pkg-config prints on the standard output.

Note Currently, all errors in pkg-config invocation results in `Nonexistent.