Module Ocb_stubblr_topkg

Hook up Ocb_stubblr with topkg.

Activating ocb-stubblr

val build_arg : Topkg.Cmd.t

build_arg is the command-line parameter that needs to be passed to ocamlbuild in order to access Ocb_stubblr inside myocamlbuild.ml.

val build_cmd : Topkg.Conf.t -> Topkg.Conf.os -> Topkg.Cmd.t

build_cmd c os is the default Topkg.build_cmd extended with build_arg.

val cmd : Topkg.Conf.t -> Topkg.Conf.os -> Topkg.fpath list -> unit Topkg.result

cmd c os files is a full invocation of build_cmd.

Install helpers

val mirage : ?xen:bool -> ?fs:bool -> Topkg.fpath -> Topkg.Pkg.install

mirage ?xen ?fs clib installs Mirage-specific multi-lib variants of the given .clib file.

xen enables mirage-xen target. Defaults to false.

fs enables mirage-freestanding target. Defaults to false.

Usage

In pkg.ml:

Making ocb-stubblr available from myocamlbuild.ml

#require "ocb-stubblr.topkg"

let build = Pkg.build ~cmd:Ocb_stubblr_topkg.cmd ()

let () = Pkg.describe ~build ...

Mirage

Pkg.describe ... @@ fun c ->
...
Ok [ ...; mirage ~xen ~fs "path/to/libstubs.clib"]